+++ src/styles/home/note/Note-styled.js
