+++ src/styles/home/note/NoteItem-styled.js
